fix: early return when request body not application/json
I found an issue where we'd bail early from parsing operations if there
was a request body that did not contain an application/json media type.
This means it was possible you could have an operation that took a
non-json request and that would result in the next operations not
generating any hurl files. The fix was just to use `while let` instead
of `if let` so the breaks worked.